refactored to use context
