+++ backend/middleware/defineCurrentUser.js
@@ -18,6 +18,7 @@ async function defineCurrentUser(req, res, next){
         }
         next()
     } catch(err){
+        req.currentUser = null
         next() 
     }
 }
