+++ frontend/src/contexts/CurrentUser.js
@@ -1,11 +1,24 @@
-import { createContext, useState } from "react";
+import { useEffect, createContext, useState } from "react";
 
 
 export const CurrentUser = createContext()
 
-function CurrentUserProvider({ children }){
+function CurrentUserProvider({ children }) {
 
     const [currentUser, setCurrentUser] = useState(null)
+
+    useEffect(() => {
+        const getLoggedInUser = async () => {
+            let response = await fetch('http://localhost:5000/authentication/profile', {
+                headers: {
+                    'Authorization': `Bearer ${localStorage.getItem('token')}`
+                }
+            })
+            let user = await response.json()
+            setCurrentUser(user)
+        }
+        getLoggedInUser()
+    }, [])
 
     return (
         <CurrentUser.Provider value={{ currentUser, setCurrentUser }}>
