+++ frontend/src/users/Login.js
@@ -1,9 +1,12 @@
-import { useState } from "react"
+import { useContext, useState } from "react"
 import { useHistory } from "react-router"
+import { CurrentUser } from "../contexts/CurrentUser"
 
 function LoginForm() {
 
     const history = useHistory()
+
+    const { setCurrentUser } = useContext(CurrentUser)
 
     const [credentials, setCredentials] = useState({
         email: '',
@@ -25,6 +28,7 @@ function LoginForm() {
         const data = await response.json()
 
         if (response.status === 200) {
+            setCurrentUser(data.user)
             localStorage.setItem('token', data.token)
             history.push(`/`)
         } else {
